Glue was running too often.
Added a separate Lambda for running the Glue update, but less frequently.  Changed the ETL Lambda so that it explicitly informs Athena of any new partitions that it finds, instead of depending on Glue to automatically find them.  That added complexity to this project but it will save ~ $30/day at this point and the costs would only increase from there.

Also made the Lambda function name(s) more descriptive since now there's more than one.

I like the idea of using this one `lambda_src` folder for all the Lambdas for now instead of setting up a separate Lambda layer.  Not as effcient as possible for the smallest-possible Lambda functions, but simpler for the humans.